+++ tools/go/tests/download_test.rs
@@ -207,6 +207,7 @@ mod go_tool {
                         version: VersionSpec::parse("1.2.0").unwrap(),
                         ..Default::default()
                     },
+                    ..Default::default()
                 })
                 .await
                 .exes
@@ -233,6 +234,7 @@ mod go_tool {
                         version: VersionSpec::parse("1.2.0").unwrap(),
                         ..Default::default()
                     },
+                    ..Default::default()
                 })
                 .await
                 .exes