+++ tools/internal-schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Updated to support proto v0.47 release.
+
 ## 0.17.0
 
 #### 🚀 Updates